pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = "ecommerce for sellers"
authors = ["therishabhmittal05 <rishabhmittal928@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
uvicorn = "^0.30.1"
pydantic = "^2.7.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
